chore: remove unused code from server (#14437)
