+++ testutil/network/util.go
@@ -3,10 +3,12 @@ package network
 import (
 	"encoding/json"
 	"fmt"
+	"net"
 	"os"
 	"path/filepath"
 	"time"
 
+	sdkerrors "cosmossdk.io/errors"
 	"github.com/tendermint/tendermint/node"
 	"github.com/tendermint/tendermint/p2p"
 	pvm "github.com/tendermint/tendermint/privval"
@@ -201,4 +203,22 @@ func writeFile(name string, dir string, contents []byte) error {
 	}
 
 	return nil
+}
+
+// Get a free address for a test tendermint server
+// protocol is either tcp, http, etc
+func FreeTCPAddr() (addr, port string, err error) {
+	l, err := net.Listen("tcp", "localhost:0")
+	if err != nil {
+		return "", "", err
+	}
+
+	if err := l.Close(); err != nil {
+		return "", "", sdkerrors.Wrap(err, "couldn't close the listener")
+	}
+
+	portI := l.Addr().(*net.TCPAddr).Port
+	port = fmt.Sprintf("%d", portI)
+	addr = fmt.Sprintf("tcp://0.0.0.0:%s", port)
+	return
 }
